refactor: controller 위치에 맞게 fixture 메서드 위치 조정
