+++ src/test/java/woowacourse/auth/acceptance/AuthAcceptanceTest.java
@@ -2,8 +2,8 @@ package woowacourse.auth.acceptance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static woowacourse.fixture.AuthFixture.findById;
-import static woowacourse.fixture.CustomerFixture.login;
+import static woowacourse.fixture.CustomerFixture.findById;
+import static woowacourse.fixture.AuthFixture.login;
 import static woowacourse.fixture.CustomerFixture.signUp;
 
 import io.restassured.response.ExtractableResponse;
