+++ src/test/java/woowacourse/fixture/AuthFixture.java
@@ -9,55 +9,16 @@ import org.springframework.http.MediaType;
 
 public class AuthFixture {
 
-    public static ExtractableResponse<Response> findById(final String token) {
-        return RestAssured
-                .given().log().all()
-                .header("Authorization", "Bearer " + token)
-                .when().get("/auth/customers/profile")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> update(final String token, final String nickname, final String password) {
+    public static ExtractableResponse<Response> login(final String userId, final String password) {
         Map<String, Object> requestBody = new HashMap<>();
-        requestBody.put("nickname", nickname);
+        requestBody.put("userId", userId);
         requestBody.put("password", password);
 
         return RestAssured
                 .given().log().all()
-                .header("Authorization", "Bearer " + token)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(requestBody)
-                .when().patch("/auth/customers/profile")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> updatePassword(final String token, final String oldPassword, final String newPassword) {
-        Map<String, Object> requestBody = new HashMap<>();
-        requestBody.put("oldPassword", oldPassword);
-        requestBody.put("newPassword", newPassword);
-
-        return RestAssured
-                .given().log().all()
-                .header("Authorization", "Bearer " + token)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(requestBody)
-                .when().patch("/auth/customers/profile/password")
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> withdraw(final String token, final String password) {
-        Map<String, Object> requestBody = new HashMap<>();
-        requestBody.put("password", password);
-
-        return RestAssured
-                .given().log().all()
-                .header("Authorization", "Bearer " + token)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(requestBody)
-                .when().delete("/auth/customers/profile")
+                .when().post("/customers/login")
                 .then().log().all()
                 .extract();
     }
